requirements.txt
flask>=3.0.0
waitress>=3.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests>=2.31.0
qrcode[pil]>=7.4.0
